requirements.txt
pandas
openpyxl
python-pptx
google-api-python-client
google-auth
google-auth-oauthlib
google-generativeai
modal
